Export mentors for a given term to Excel file
Allows for exporting of mentors for a given term to an excel file, which lists
relevant risk management and enrollement data along with contact information
for all mentors. Includes a number of convienience methods for retrieving
this information.

Also includes a rewrite of the JavaScript used to update report generation
state to use jQuery instead of Prototype.