Fixes bug with year-edge activity log start dates
I was using Date#commercial to calculate week numbers after date
calculation. Apparently that method was always returning the last week
of 2013 for a date that spanned across the 2014 year-edge boundary
because of the way it does date math. I’ve always been a bit confused
by that method. So I replaced
ActivityLog#find_or_create_by_mentor_id_and_week_and_year with
#find_by_mentor_id_and_date, and then calculate the week beginning and
end from inside that method instead of relying on the downstream input.
This actually creates a cleaner implementation and seems to work very
well.

Important note is that ActivityLog stores the start and end date in the
database, so when those rogue logs were created with the bad 2013 date,
it was stored in the database that way. We’ll need to go through and
manually clean those up after this fix is pushed.