Adds spinner when saving attendance
Also, the updated method compares the expected value with the returned
value from the server, so that the spinner keeps spinning until all
save steps have completed. Because a user can click multiple times to
cycle through the attendance options, previously this could result in
the final result not being saved but appearing as thought it had.