Adds dynamic tagging functionality system-wide
Any model can be made taggable by adding `acts_as_taggable` in the
model. Then in a form, add a select box with class `taggable` to allow
tags to be added. By default the options are the collection value for
the select, but by adding `”data-tags” => true` to the input_html, it
will allow any arbitrary tag to be set.