Make some fields disabled for participants
This sets up a new method in User called `can_edit_own?` that is used
to check for permission to change a particular attribute. This method
is versatile and can be used in other places in the future. Changes
here also update the `profile` and `participant_signup` controllers to
manually check for these controlled fields and trash the changes if
necessary.