Configure filters to warn on false
Previously all filters resulted in a green check or a red X, but that
makes it hard to identify which items are just informational and which
need action taken. This commit adds the ability to specify which
filters should trigger a warning, and updates the views to accommodate
this change.