Move participant tab content under respective controllers
Previously all tab content on `participant#show` were populated with
partials in the `/views/participants` directory. Now tab content is
populated with the appropriate `index.html.*` file from the
corresponding child resource.

Also introduces a global controller listener that watches for presence
of the `X-DreamSIS-Render-Intent` header, which can be set to
`no-layout` to automatically prevent the app layout being rendered in
the response. This is useful for the participant tabs because the
client can request the tab content using the same controller method.